refactor: move SurfaceType to Game/Types
