+++ frontend/src/game/levels/types/index.ts
@@ -1,7 +1,6 @@
 import LevelId from './levelIds';
 import LevelLoad from './levelLoad';
-import SurfaceType from './surfaceType';
 
 export {
-  LevelId, LevelLoad, SurfaceType,
+  LevelId, LevelLoad,
 };
