+++ frontend/src/game/types/index.ts
@@ -1,0 +1,3 @@
+import SurfaceType from './surfaceType';
+
+export default SurfaceType;
