+++ frontend/src/game/types/surfaceType.ts
